Extend pykat.kat.beamTrace for non 1um wavelengh
pykat.kat.beamTrace was defaulting to 1.064um
wavelenght when producing a beam trace.

This was likley a bug. I have updated the
code so the wavelength is preserved. Some results
for a simple telescope match my analytics
and also match a Finesse simulation with a
`bp` detector placed at several nodes.

Tested at 2um wavelenght.